requirements.txt
numpy
torch
torchvision
efficientnet-pytorch
Pillow
mediapipe
opencv-python